split off compute
