+++ benchmark_flops/compute.py
@@ -1,0 +1,16 @@
+import numpy as np
+import time
+
+def compute_flops(loopcount, MAT_N):
+    
+    A = np.arange(MAT_N**2, dtype=np.float64).reshape(MAT_N, MAT_N)
+    B = np.arange(MAT_N**2, dtype=np.float64).reshape(MAT_N, MAT_N)
+
+    t1 = time.time()
+    for i in range(loopcount):
+        c = np.sum(np.dot(A, B))
+
+    FLOPS = 2 *  MAT_N**3 * loopcount
+    t2 = time.time()
+    return FLOPS / (t2-t1)
+
